backend/requirements.txt
fastapi==0.115.2
uvicorn==0.30.6
SQLAlchemy==2.0.36
pymysql==1.1.1
pydantic==2.9.2
python-dotenv==1.0.1
